update: get categories data
